okx_bot/requirements.txt
python-telegram-bot==20.7
apscheduler>=3.10.4
requests>=2.31.0
websockets>=11.0.3
okx>=2.1.1
